backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
google-auth
google-auth-oauthlib
google-api-python-client
joblib
scikit-learn